Add files via upload
With the increase in the number of news it has got difficult for users to access news of his/her interest which makes it a necessity to categorize a  news so that it could be easily accessed. Categorization refers to grouping that allows easier navigation among articles. Internet news needs to be divided into categories. This will help users to access the news of their interest in real-time without wasting any time. When it comes to news it is much difficult to classify as news are continuously appearing that need to be processed and those news could be never-seen-before and could fall in a new category


Text classification is also known as text tagging or text categorization is the process of categorizing text into organized groups. By using NLP text classifier can automatically analyse the text and then they assign a set of predefined tags or predefined categories based on its contents or based on its textual contents. 
Unstructured text is everywhere such as emails, chat conversation, websites and social media but it is very hard to extract value from this data unless it is organized in a certain way, doing so in earlier time, used to be difficult task and expensive as well.
Since it requires spending lot of time and resources to manually solve the data or creating hand crafted rules and features required are difficult to maintain.

Now text classifiers with NLP have proven to be great alternative to structure textual data in a fast, efficient, cost effective and a scalable way.  
Text classification is becoming an increasingly important part of many businesses as it allows to easily get insights of data and automate various business processes.

Thus the goal will be to build a system that can accurately classify previously unseen news articles into the right category. Text classification is the process of assigning tags or categories to text according to its content. It's one of the fundamental tasks in Natural Language Processing (NLP) with broad applications such as sentiment analysis, topic labeling, spam detection, and intent detection, Tagging Products/Contents, Differentiate between legitimate messages and spam mails, News Portals.

Text classification process involves following main steps for classification of news article. These steps are data collection, pre-processing, feature selection, classification techniques application, and evaluating performance measures. As the first step of news classification is accumulating news from various sources.., we are having a dataframe of BBC News classification.